Break up cli.py
