+++ netkan/netkan/cli/restore_status.py
@@ -1,0 +1,13 @@
+import click
+
+from ..status import ModStatus
+
+
+@click.command()
+@click.argument('filename')
+def restore_status(filename):
+    click.echo(
+        'To keep within free tier rate limits, this could take some time'
+    )
+    ModStatus.restore_status(filename)
+    click.echo('Done!')
